+++ src/DotNetCore.Collections.Paginable.DosOrm/DotNetCore/Collections/Paginable/Extensions/SolidPageExtensions.cs
@@ -1,100 +1,21 @@
-﻿using System;
 using Dos.ORM;
-using DotNetCore.Collections.Paginable.Internal;
 
 // ReSharper disable once CheckNamespace
 namespace DotNetCore.Collections.Paginable
 {
-    /// <summary>
-    /// Extensions for solid page for DosORM
-    /// </summary>
-    public static class SolidPageExtensions
+    [SolidPageExtensionsFor(
+        ormName: "DosORM",
+        sourceTypeFullName: "Dos.ORM.FromSection<T>",
+        sourceParamName: "query",
+        pageTypeName: "DosPage",
+        paginableTypeName: "PaginableDosQuery",
+        factoryTypeName: "PaginableDosCollFactory",
+        helperTypeName: "DosHelper",
+        GenericConstraint = "Entity",
+        HasAdditionalQueryFunc = true,
+        ToPaginableHasAdditionalQueryFunc = true,
+        SourceDescriptor = "DosQueryable")]
+    public static partial class SolidPageExtensions
     {
-        /// <summary>
-        /// Make original DosQueryable result to DosPage collection.
-        /// </summary>
-        /// <typeparam name="T">element type of your enumerable result</typeparam>
-        /// <param name="query">DosQueryable</param>
-        /// <param name="limitedMemberCount">limited member count</param>
-        /// <param name="additionalQueryFunc"></param>
-        /// <returns></returns>
-        /// <example>
-        /// <code>
-        /// var paginable = query.ToPaginable();
-        /// var page = paginable.GetPage(15);
-        /// </code>
-        /// </example>
-        public static PaginableDosQuery<T> ToPaginable<T>(this FromSection<T> query, int? limitedMemberCount = null,
-            Func<FromSection<T>, FromSection<T>>? additionalQueryFunc = null) where T : Entity
-            => PaginableDosCollFactory.CreatePageSet(query, limitedMemberCount: limitedMemberCount, additionalQueryFunc: additionalQueryFunc);
-
-        /// <summary>
-        /// Make original DosQueryable result to DosPage collection.
-        /// </summary>
-        /// <typeparam name="T">element type of your enumerable result</typeparam>
-        /// <param name="query">DosQueryable</param>
-        /// <param name="pageSize">page size</param>
-        /// <param name="limitedMemberCount">limited member count</param>
-        /// <param name="additionalQueryFunc"></param>
-        /// <returns></returns>
-        /// <example>
-        /// <code>
-        /// var paginable = query.ToPaginable();
-        /// var page = paginable.GetPage(15);
-        /// </code>
-        /// </example>
-        public static PaginableDosQuery<T> ToPaginable<T>(this FromSection<T> query, int pageSize, int? limitedMemberCount = null,
-            Func<FromSection<T>, FromSection<T>>? additionalQueryFunc = null) where T : Entity
-            => PaginableDosCollFactory.CreatePageSet(query, pageSize, limitedMemberCount, additionalQueryFunc: additionalQueryFunc);
-
-        /// <summary>
-        /// Get specific page from original DosQueryable source
-        /// </summary>
-        /// <typeparam name="T">element type of your DosQueryable source</typeparam>
-        /// <param name="query">original DosQueryable source</param>
-        /// <param name="pageNumber">page number</param>
-        /// <param name="additionalQueryFunc"></param>
-        /// <returns></returns>
-        /// <example>
-        /// <code>
-        /// var page = query.GetPage(15);
-        /// var totalMemberCount = page.TotalMemberCount;
-        /// </code>
-        /// </example>
-        public static IPage<T> GetPage<T>(this FromSection<T> query, int pageNumber, Func<FromSection<T>, FromSection<T>>? additionalQueryFunc = null) where T : Entity
-            => GetPage(query, pageNumber, PaginableSettingsManager.Settings.DefaultPageSize, additionalQueryFunc: additionalQueryFunc);
-
-        /// <summary>
-        /// Get specific page from original DosQueryable source
-        /// </summary>
-        /// <typeparam name="T">element type of your DosQueryable source</typeparam>
-        /// <param name="query">original DosQueryable source</param>
-        /// <param name="pageNumber">page number</param>
-        /// <param name="pageSize">page size</param>
-        /// <param name="additionalQueryFunc"></param>
-        /// <returns></returns>
-        /// <exception cref="ArgumentNullException"><paramref name="query"/> is <c>null</c>.</exception>
-        /// <exception cref="ArgumentOutOfRangeException">
-        /// <paramref name="pageNumber"/> is less than one, or <paramref name="pageSize"/> is less than one.
-        /// </exception>
-        /// <example>
-        /// <code>
-        /// var page = paginable.GetPage(15);
-        /// </code>
-        /// </example>
-        public static IPage<T> GetPage<T>(this FromSection<T> query, int pageNumber, int pageSize, Func<FromSection<T>, FromSection<T>>? additionalQueryFunc = null)
-            where T : Entity
-        {
-            if (query is null)
-                throw new ArgumentNullException(nameof(query), $"{nameof(query)} can not be null.");
-
-            if (pageNumber < 1)
-                throw new ArgumentOutOfRangeException(nameof(pageNumber), $"{nameof(pageNumber)} can not be less than one");
-
-            if (pageSize < 1)
-                throw new ArgumentOutOfRangeException(nameof(pageSize), $"{nameof(pageSize)} can not be less than one");
-
-            return new DosPage<T>(query, pageNumber, pageSize, DosHelper.Count(query), additionalQueryFunc: additionalQueryFunc);
-        }
     }
 }